Extract Item.Print
